Add keybindings for remotes.
At least some remotes masquerade as keyboards, and trigger events that
Keynote and/or PowerPoint expect. Epresent doesn’t generally understand
these.

This adds PgUp/PgDn for prev/next page, as well as F5 in org-mode to
trigger `epresent-run`.

Also made it so calling `epresent-run` when already in `epresent-mode`
is a no-op (easy to hit the wrong button on the remote, and better to
not get a backtrace when that happens).