Adjust frame text size, rather than buffer.
This allows the mode-line to scale along with the text. It also means we
only have to set the size when we create the frame, and it doesn’t
affect any other windows containing the org buffer.

This also switches from a relative scale to an absolute size (in 1/10
pt). I was having trouble figuring out what the scaling should be, and
also don’t see any connection between the size I like to edit and the
size that should be presented to an audience. I set the default to 50
pt, since that seems to be a common default in various presentation software.