Replace text-size with face-attributes.
Fixes #48, #50, #51, and #53.

This allows much more fine-grained control over epresent styling. In
addition to being able to CUSTOMIZE faces it also adds a
`#+EPRESENT-FACE-ATTRIBUTES` option to Org that sets faces for a
specific presentation.

Here is an example from a recent presentation I gave:

```
```

It also removes epresent-specific face settings.