Encapsulate presentation-specific values.
This introduces a class `epresent-presentation` that holds anything
specific to a particular presentation. It also makes things much more
lexically-scoped (although there is still too much that depends on
`epresent--default-presentation`).

This is a first step toward two features:

1. running multiple presentations simultaneously and
2. having a 2nd frame for a presentation, displaying speaker notes, etc.

The one “breaking” change here is that the `epresent-mode-line` variable
has changed. It’s now a function (taking an `epresent-presentation`
instance) that evaluates to a mode-line format.